requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-extensions==3.0.6
djangorestframework==3.12.4
gunicorn==19.9.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0